ファイルアップロード 実装

+++ functions/admin/users/deleteResume.js
@@ -5,20 +5,23 @@ const db = require("../../firebase").db;
 const location = require("../../firebase").location;
 const runtime = require("../../firebase").runtime;
 
-const userAuthenticated =
-  require("./functions/userAuthenticated").userAuthenticated;
-
 const dataTime = Date.now();
 
 exports.deleteResume = functions
   .region(location)
   .runWith(runtime)
   .https.onCall(async (data, context) => {
-    await userAuthenticated({ context: context, demo: true });
+    if (context.auth.uid !== functions.config().admin.uid) {
+      throw new functions.https.HttpsError(
+        "cancelled",
+        "無効なユーザーのため、処理中止",
+        "firebase"
+      );
+    }
 
     await db
       .collection("persons")
-      .doc(context.auth.uid)
+      .doc(data.uid)
       .get()
       .then(async (doc) => {
         doc.exists && deleteFile(doc);
